src/requirements.txt
bcrypt==3.2.0
cffi==1.14.3
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==3.25.0
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.9.0
marshmallow-sqlalchemy==0.24.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.15.0
six==1.15.0
SQLAlchemy==1.3.20
Werkzeug==1.0.1
